lesson 18 class changes profiletests create and move profile name change
