+++ src/test/java/Homework18.java
@@ -1,0 +1,20 @@
+import org.testng.annotations.Test;
+
+public class Homework18 extends BaseTest {
+
+    @Test
+    public void testPlaySong() {
+
+
+
+    }
+
+public void playSong() {
+
+}
+
+public void isSonPlaying () {
+
+}
+
+}
