+++ src/test/java/Homework19.java
@@ -1,0 +1,2 @@
+public class Homework19 {
+}
