+++ src/test/java/LoginTests.java
@@ -5,6 +5,8 @@ import org.testng.annotations.Test;
 
 public class LoginTests extends BaseTest {
 
+
+
     @Test
     public static void loginEmptyEmailPasswordTest () {
 
@@ -13,27 +15,28 @@ public class LoginTests extends BaseTest {
         Assert.assertEquals(driver.getCurrentUrl(), url);
         driver.quit();
     }
-    @Test
-    public static void updateProfileNameTest () throws InterruptedException {
-
-        navigateToPage();
-        provideEmail("demo@class.com");
-        providePassword("te$t$tudent");
-        clickSubmit();
-
-        Thread.sleep(2000);
-        clickavatarIcon();
-
-        String randomName = generateRandomName();
-
-            provideCurrentPassword("te$t$tudent");
-            provideProfileName(randomName);
-            clickSaveButton();
-
-        Thread.sleep(2000);
-            WebElement actualProfileName = driver.findElement(By.cssSelector("a.view-profile>span"));
-            Assert.assertEquals(actualProfileName.getText(), randomName);
-}
+//moving this to ProfileTests
+//    @Test
+//    public static void updateProfileNameTest () throws InterruptedException {
+//
+//        navigateToPage();
+//        provideEmail("demo@class.com");
+//        providePassword("te$t$tudent");
+//        clickSubmit();
+//
+//        Thread.sleep(2000);
+//        clickavatarIcon();
+//
+//        String randomName = generateRandomName();
+//
+//            provideCurrentPassword("te$t$tudent");
+//            provideProfileName(randomName);
+//            clickSaveButton();
+//
+//        Thread.sleep(2000);
+//            WebElement actualProfileName = driver.findElement(By.cssSelector("a.view-profile>span"));
+//            Assert.assertEquals(actualProfileName.getText(), randomName);
+//}
 
 }
